+++ boundlexx/boundless/utils.py
@@ -198,15 +198,19 @@ def html_name(string, strip=False, colors=None):
             )
         # replace emoji with resolved emoji
         else:
+            user_name = inner.lower()
             try:
-                emoji = Emoji.objects.get_by_name(inner.lower())
+                emoji = Emoji.objects.get_by_name(user_name)
             except Emoji.DoesNotExist:
                 pass
             else:
                 if strip:
                     final_string = final_string.replace(format_string, inner, 1)
                 else:
-                    html_emoji = f'<img src="{emoji.image.url}" class="emoji">'
+                    html_emoji = (
+                        f'<img src="{emoji.image.url}" class="emoji"'
+                        f' alt="emoji {user_name}" title=":{user_name}:">'
+                    )
                     final_string = final_string.replace(format_string, html_emoji, 1)
 
     return mark_safe(final_string)  # nosec